change factory namespace
